move NavTitle to NavBarNew
